requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
apscheduler==3.10.4
telethon==1.34.0
vk-api==11.9.9
aiosqlite==0.19.0
